Add files via upload
was able to finish the method areAnagrams() but the problem it is having is that it only looks if the second string contains the letter found in the first string, not comparing the exact number of letters each word has. Does that make sense? If you can find a way to fix that let me know. If this still does not make sense, refer to the JUnit test I posted.